% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_event_data.R
\name{create_event_data}
\alias{create_event_data}
\title{Create Event Data}
\usage{
create_event_data(
  maindata,
  timevar,
  unitvar,
  cohortvar,
  control_group,
  base_time = -1,
  anycohortvar = NULL,
  covariate_base_stratify = NULL,
  covariate_base_balance = NULL,
  covariate_base_balance_linear = NULL,
  covariate_base_balance_linear_subset = NULL,
  balanced_panel = TRUE,
  check_not_treated = FALSE,
  stratify_by_cohort = FALSE,
  lower_event_time = -Inf,
  upper_event_time = Inf,
  min_control_gap = NULL,
  max_control_gap = NULL,
  treat_criteria = NULL,
  base_restrict = NULL,
  base_restrict_treated = NULL,
  verbose = TRUE,
  copy_dataset = TRUE,
  validate = TRUE,
  combine = FALSE
)
}
\arguments{
\item{maindata}{A data.table containing the main data.}

\item{timevar}{The name of the column representing time.}

\item{unitvar}{The name of the column representing unit (e.g., household).}

\item{cohortvar}{The name of the column representing the period when the event of interest arises.}

\item{control_group}{Options: "both", "never", and "later".}

\item{base_time}{The base time for analysis (default is -1).}

\item{anycohortvar}{The name of the column representing the period when ANY kind of event arises, used to find control cohorts (default is NULL).}

\item{covariate_base_stratify}{Vector of variable names; treatment effects are stratified by their joint unique values (default is NULL).}

\item{covariate_base_balance}{Vector of variable names to include in finding exact matches for treated units (default is NULL).}

\item{covariate_base_balance_linear}{Vector of variable names to include linearly in propensity score estimator (default is NULL).}

\item{covariate_base_balance_linear_subset}{Vector of variable names within which the linear propensity score regression will be estimated (default is NULL).}

\item{balanced_panel}{Logical, whether to enforce a balanced panel (default is TRUE).}

\item{check_not_treated}{If TRUE, only allows controls to be used for a cohort if they are observed in the dataset to be untreated in that cohort (default is FALSE).}

\item{stratify_by_cohort}{Logical, whether to stratify by cohort (default is FALSE).}

\item{lower_event_time}{Earliest period (relative to treatment time) for which to estimate effects (default is -Inf).}

\item{upper_event_time}{Latest period (relative to treatment time) for which to estimate effects (default is Inf).}

\item{min_control_gap}{Controls must be treated at least this many years away from matched treated units (default is NULL).}

\item{max_control_gap}{Controls must be treated no more than this many years away from matched treated units (default is NULL).}

\item{treat_criteria}{Replacement of onsetagevar, provide a character column that must == TRUE for the treated (default is NULL).}

\item{base_restrict}{A single variable, restricting to households for which the variable == 1 in the base period (default is NULL).}

\item{base_restrict_treated}{A single variable, restricting TREATED GUYS to households for which the variable == 1 in the base period (default is NULL).}

\item{verbose}{Logical, whether to display messages (default is TRUE).}

\item{copy_dataset}{Logical, whether to create a copy of the dataset to avoid changing the original input (default is TRUE).}

\item{validate}{Logical, whether to validate the data (default is TRUE).}

\item{combine}{Logical, whether to combine the resulting event data or return a list (default is TRUE).}
}
\value{
Event data as a data.table.
}
\description{
Generates event data for treated households based on the input parameters.
}
\details{
This function generates event data for treated households based on the provided input parameters.
}
\examples{

create_event_data(maindata, "time", "id", "cohort", "both")


}
