% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_confound_event.R
\name{diagnose_confound_event}
\alias{diagnose_confound_event}
\title{Diagnose Event Confoundedness}
\usage{
diagnose_confound_event(
  data,
  timevar,
  cohortvar,
  cohortvar2,
  control_option = "both",
  weightvar = NA,
  anticipation = 0
)
}
\arguments{
\item{data}{A data.table containing the panel data.}

\item{timevar}{The name of the time variable.}

\item{cohortvar}{The name of the cohort (group) variable.}

\item{cohortvar2}{The name of the second cohort (group) variable.}

\item{control_option}{The control units used for the DiD estimates. Default is "both".}

\item{weightvar}{The name of the weight variable, if not specified will cluster on unit level (optional).}

\item{anticipation}{Period of anticipation}
}
\value{
A data.table containing the contamination ratio for each group-time.
}
\description{
Diagnose Event Confoundedness
}
\examples{
# simulated data
simdt <- sim_did(1e+03, 10, cov = "cont", second_cov = TRUE, 
second_outcome = TRUE, second_cohort = TRUE)
dt <- simdt$dt

diagnose_confound_event(dt, "time", "G", "G2")

}
\keyword{computation}
\keyword{data}
\keyword{did}
\keyword{difference-in-differences}
\keyword{estimation}
\keyword{fast}
\keyword{panel}
