% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_did_dynamics}
\alias{plot_did_dynamics}
\title{Create plot for Difference-in-Differences (DiD) analysis.}
\usage{
plot_did_dynamics(x, margin = "event_time")
}
\arguments{
\item{x}{A data table containing the results of the DiD analysis. It should include columns for 'att' (average treatment effect), 'se' (standard error), and 'event_time' (time points).}

\item{margin}{the x-axis of the plot}
}
\value{
A ggplot2 object representing the event study plot.
}
\description{
This function generates an plot based on the results of a DiD analysis.
}
