% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_did.R
\name{sim_did}
\alias{sim_did}
\title{Simulate a Difference-in-Differences (DiD) dataset}
\usage{
sim_did(
  sample_size,
  time_period,
  untreated_prop = 0.3,
  cov = "no",
  hetero = "all",
  second_outcome = FALSE,
  na = "none",
  balanced = TRUE,
  seed = NA,
  stratify = FALSE,
  treatment_assign = "latent"
)
}
\arguments{
\item{sample_size}{The number of units in the dataset.}

\item{time_period}{The number of time periods in the dataset.}

\item{untreated_prop}{The proportion of units that remain untreated.}

\item{cov}{The type of covariate to include ("int" for integer, "no" for constant, "cont" for continuous).}

\item{hetero}{The heterogeneity type for treatment effects ("all" or "dynamic").}

\item{second_outcome}{Should a second outcome variable be generated?}

\item{na}{The type of missing data to introduce ("none", "x", "y", or "both").}

\item{balanced}{Should the dataset be balanced (whether some unit have missing observation in some period)}

\item{seed}{The random seed for reproducibility.}

\item{stratify}{Should treatment effect be stratified?}

\item{treatment_assign}{The method of treatment assignment ("latent" or "uniform").}
}
\value{
A list containing the generated dataset (dt) and treatment effect information (att).
}
\description{
This function generates a synthetic dataset for conducting a Difference-in-Differences (DiD) analysis.
}
\examples{
sim_data <- sim_did(sample_size = 100, time_period = 5)

}
