% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastdid.R
\name{fastdid}
\alias{fastdid}
\title{Fast DID Estimation}
\usage{
fastdid(
  dt,
  timevar,
  cohortvar,
  unitvar,
  outcomevar,
  control_option = "both",
  result_type = "group_time",
  boot = FALSE,
  biters = 1000,
  weightvar = NULL,
  clustervar = NULL,
  covariatesvar = NULL,
  copy = FALSE,
  validate = TRUE
)
}
\arguments{
\item{dt}{A data table containing the panel data.}

\item{timevar}{The name of the time variable.}

\item{cohortvar}{The name of the cohort (group) variable.}

\item{unitvar}{The name of the unit (id) variable.}

\item{outcomevar}{The name of the outcome variable.}

\item{control_option}{A character string indicating the control option for the DID estimation. Default is "both".}

\item{result_type}{A character string indicating the type of result to be returned. Default is "group_time".}

\item{boot}{Logical, indicating whether bootstrapping should be performed. Default is FALSE.}

\item{biters}{The number of bootstrap iterations. Only relevant if boot = TRUE. Default is 1000.}

\item{weightvar}{The name of the weight variable (optional).}

\item{clustervar}{The name of the cluster variable (optional).}

\item{covariatesvar}{A character vector containing the names of covariate variables (optional).}
}
\value{
A data table containing the estimated treatment effects and standard errors.
}
\description{
This function performs Difference-in-Differences (DID) estimation using fast computation techniques.
}
\examples{
# Example usage of fastdid function
result <- fastdid(data, "time", "cohort", "unit")

}
\seealso{
\code{\link{estimate_gtatt}}, \code{\link{aggregate_gt_result}}, \code{\link{get_se}}, \code{\link{convert_targets}}
}
\keyword{computation}
\keyword{data}
\keyword{difference-in-differences}
\keyword{estimation}
\keyword{fast}
\keyword{panel}
