% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_event_dynamics.R
\name{plot_did_dynamics}
\alias{plot_did_dynamics}
\title{Create an event study plot for Difference-in-Differences (DiD) analysis.}
\usage{
plot_did_dynamics(dt)
}
\arguments{
\item{dt}{A data table containing the results of the DiD analysis. It should include columns for 'att' (average treatment effect), 'se' (standard error), and 'event_time' (time points).}
}
\value{
A ggplot2 object representing the event study plot.
}
\description{
This function generates an event study plot based on the results of a DiD analysis.
}
